Tyler: This is what I have for the database portion. I copied these over from a local project named Class_Project_Database_Portion. You all will see that reflected in the namespace(s). Feel free to change it.
I included the appsettings.json and Program.cs so everybody could see how it is setup specifically for the database.

Also, I included the index for Andy so that he has workable data for the view.

Even if we do not use MVC, if Andy has MVC starting out, he should be able to use this to begin working and Chris should be able to use this to integrate over to APIs.

The dependencies I have are:

Microsoft.Data.SqlClient
Microsoft.EntityFrameworkCore.SqlServer
Microsoft.EntityFrameworkCore.Tools
Dapper          - version 2.1.66